Check if KP is enabled before processing order/cart
Prevent KP from running if it is not enabled. This can happen if the plugin is activated, but the payment method is not enabled which may result in undefined index errors.

This change include check for redundancy so that we don't check if the plugin is enabled more than necessary.